% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/india_air.R
\docType{data}
\name{india_air}
\alias{india_air}
\title{Air pollutant measures in six Indian cities}
\format{
A dataframe with 12051 rows and 14 variables:
\describe{
  \item{city}{City the measurement was taken in (character) (Delhi, Hyderabad, Lucknow, Bengluru, Chennai, Ahmedabad, Mumbai)}
  \item{date}{Y-M-D date variable for the date (lubridate date)}
  \item{PM2.5}{Concentration of particulate matter less than 2.5 micrometers in diameter (numeric), micrograms per meter cubed}
  \item{NO}{Nitrogen monoxide concentration (numeric), micrograms per meter cubed}
  \item{NO2}{Nitrogen Dioxide concentration (numeric), micrograms per meter cubed}
  \item{NOx}{Ntirous oxides concentration (numeric), micrograms per meter cubed}
  \item{CO}{Carbon monoxide concentration (numeric), milligrams per meter cubed}
  \item{SO2}{Sulfur dioxide concentration (numeric), micrograms per meter cubed}
  \item{O3}{Ozone concentration (numeric), micrograms per meter cubed}
  \item{benzene}{Benzene concentration (numeric), micrograms per meter cubed}
  \item{toluene}{Toluene concentration (numeric), micrograms per meter cubed}
  \item{xylene}{Xylene concentration (numeric), micrograms per meter cubed}
  \item{AQI}{Air Quality Index (numeric)}
  \item{AQI_bucket}{Description of AQI(character)(Severe, Very Poor, Poor, Moderate, Satisfactory, Good)}
}
}
\source{
{National Air Quality Standards in India as of 2009} \url{https://cpcb.nic.in/uploads/National_Ambient_Air_Quality_Standards.pdf}

{National Air Quality Index Definitions} \url{https://cpcb.nic.in/displaypdf.php?id=bmF0aW9uYWwtYWlyLXF1YWxpdHktaW5kZXgvQWJvdXRfQVFJLnBkZg==}

{List of CAAQM Stations Installed Across India} \url{https://app.cpcbccr.com/caaqms/download?filename=Station_List.xlsx}

{National Air Quality Index Report} \url{https://app.cpcbccr.com/ccr_docs/FINAL-REPORT_AQI_.pdf}
}
\usage{
india_air
}
\description{
A variety of air pollution measurements by city and date across India for the years of 2015-2020. Not all measurements were collected in all cities on all days.
}
\keyword{datasets}
